migr. corregidas claves foraneas

+++ database/migrations/2021_10_29_052543_create_t_adelantos_table.php
@@ -17,7 +17,7 @@ class CreateTAdelantosTable extends Migration
             $table->id();
             $table->integer('monto');
             $table->timestamps();
-            $table->integer('id_vendedor');
+            $table->unsignedBigInteger('id_vendedor');
             $table->foreign('id_vendedor')->references('id')->on('t_vendedores');
         });
     }
