+++ database/migrations/2021_10_29_215235_create_t_cuentas_table.php
@@ -28,13 +28,13 @@ class CreateTCuentasTable extends Migration
             $table->string('Vencimiento');
             $table->integer('N_Cuotas');
             $table->string('Situacion');
-            $table->integer('id_cliente');
+            $table->unsignedBigInteger('id_cliente');
             $table->foreign('id_cliente')->references('id')->on('t_clientes');
-            $table->integer('id_producto');
+            $table->unsignedBigInteger('id_producto');
             $table->foreign('id_producto')->references('id')->on('t_productos');
-            $table->integer('id_vendedor');
+            $table->unsignedBigInteger('id_vendedor');
             $table->foreign('id_vendedor')->references('id')->on('t_vendedores');
-            $table->integer('id_cobrador');
+            $table->unsignedBigInteger('id_cobrador');
             $table->foreign('id_cobrador')->references('id')->on('t_cobradores');
             $table->timestamps();
         });
