+++ database/migrations/2021_10_29_215236_create_t_licitaciones_table.php
@@ -17,7 +17,7 @@ class CreateTLicitacionesTable extends Migration
             $table->id();
             $table->integer('monto');
             $table->timestamps();
-            $table->integer('id_cuenta');
+            $table->unsignedBigInteger('id_cuenta');
             $table->foreign('id_cuenta')->references('id')->on('t_cuentas');
         });
     }
